+++ routes/api/contacts-routes.js
@@ -1,0 +1,17 @@
+const express = require("express");
+
+const ctrl = require("../../controllers/contacts-controllers");
+
+const router = express.Router();
+
+router.get("/", ctrl.getAllContacts);
+
+router.get("/:contactId", ctrl.getContactById);
+
+router.post("/", ctrl.addContact);
+
+router.delete("/:id", ctrl.deleteContactById);
+
+router.put("/:id", ctrl.updateContactById);
+
+module.exports = router;
